Move TotalPowerAndTotalRolls to power.calculator and mark helper classes as package only
